Finished solver for Simon Tatham's tents... sort of
It's very fragile and only works on my resolution. Couldn't figure out
how to reliably determine board size and particularly how to read the
numbers on the sides. Currently I have hardcoded the number of black
pixels in the glyphs at a particular resolution, so not ideal.

But the algorithm itself works well!